save 5.2 修改记录编辑、复习页面及逻辑

+++ app/src/main/java/com/wuruoye/know/ui/edit/vm/IReviewStrategyEditVM.kt
@@ -12,10 +12,10 @@ interface IReviewStrategyEditVM {
     val reviewStrategy: MutableLiveData<ReviewStrategy>
     val submitResult: MutableLiveData<Result<ReviewStrategy>>
     val remTime: Int
-    val gapTime: Long
+    val gapTime: Int
 
     fun saveReviewStrategy(title: String)
     fun setReviewStrategyId(id: Long)
     fun setRemTime(time: Int)
-    fun setGapTime(time: Long)
+    fun setGapTime(time: Int)
 }
